requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
python-dotenv==0.21.1
httpx==0.24.1
pydantic==2.4.2
python-multipart==0.0.6 